completed form to create sales record and list of sales, need to add functionality in create sales record form to recognize sold vehicles

+++ sales/api/sales_rest/models.py
@@ -43,4 +43,4 @@ class SalesRecord(models.Model):
         related_name="salesrecords",
         on_delete=models.CASCADE
     )
-    price = models.PositiveBigIntegerField()
+    price = models.IntegerField()
